Extract schema normalisation helpers into schema.rs module
Create new src/tools/registry/schema.rs containing:
- parse_schema_string: Parse and validate JSON schema strings from backends
- normalized_schema: Normalise schema values for WASM tool registration
- Comprehensive rstest tests for both functions

Update src/tools/registry.rs:
- Add mod schema declaration

Update src/tools/registry/loader.rs:
- Import normalized_schema from schema module
- Remove duplicate function definitions (90 lines removed)
- Remove duplicate test module

This extraction improves code organisation by consolidating schema
normalisation logic in a dedicated module, reducing loader.rs from
545 to 455 lines.
